+++ Commands/ProfessionCommands.cs
@@ -1,6 +1,5 @@
 using Bloodcraft.Services;
 using Bloodcraft.Systems.Professions;
-using ProjectM;
 using ProjectM.Network;
 using Stunlock.Core;
 using Unity.Entities;
@@ -9,12 +8,6 @@ using VampireCommandFramework;
 namespace Bloodcraft.Commands;
 internal static class ProfessionCommands
 {
-    static SetDebugSettingEvent BuildingCostsDebugSetting = new()
-    {
-        SettingType = DebugSettingType.BuildCostsDisabled, // Assuming this is the correct DebugSettingType for building costs
-        Value = false
-    };
-
     [Command(name: "logProfessionProgress", shortHand: "log p", adminOnly: false, usage: ".log p", description: "Toggles profession progress logging.")]
     public static void LogProgessionCommand(ChatCommandContext ctx)
     {
@@ -97,6 +90,7 @@ internal static class ProfessionCommands
 
         LocalizationService.HandleReply(ctx, $"{professionHandler.GetProfessionName()} set to [<color=white>{level}</color>] for <color=green>{foundUser.CharacterName}</color>");
     }
+
     [Command(name: "listProfessions", shortHand: "lp", adminOnly: false, usage: ".lp", description: "Lists professions available.")]
     public static void ListProfessionsCommand(ChatCommandContext ctx)
     {
@@ -107,20 +101,5 @@ internal static class ProfessionCommands
         }
         string professions = ProfessionHandlerFactory.GetAllProfessions();
         LocalizationService.HandleReply(ctx, $"Available professions: {professions}");
-    }
-
-
-    [Command(name: "toggleBuildingCosts", shortHand: "tbc", adminOnly: true, usage: ".tbc", description: "Toggles building costs, useful for setting up a castle linked to your heart easily.")]
-    public static void ToggleBuildingCostsCommand(ChatCommandContext ctx)
-    {
-        User user = ctx.Event.User;
-
-        DebugEventsSystem existingSystem = Core.DebugEventsSystem;
-
-        BuildingCostsDebugSetting.Value = !BuildingCostsDebugSetting.Value;
-        existingSystem.SetDebugSetting(user.Index, ref BuildingCostsDebugSetting);
-
-        string toggleColor = BuildingCostsDebugSetting.Value ? "<color=green>enabled</color>" : "<color=red>disabled</color>";
-        LocalizationService.HandleReply(ctx, $"Building costs removed: {toggleColor}");
     }
 }