+++ Patches/DeathEventSystemPatch.cs
@@ -10,7 +10,6 @@ using ProjectM.Network;
 using Stunlock.Core;
 using Unity.Collections;
 using Unity.Entities;
-using static Il2CppSystem.Data.Common.ObjectStorage;
 
 namespace Bloodcraft.Patches;