+++ Systems/Professions/ProfessionInterface.cs
@@ -1,9 +1,7 @@
 ﻿using Stunlock.Core;
-using Unity.Jobs;
 
 namespace Bloodcraft.Systems.Professions
 {
-    
     public interface IProfessionHandler 
     {
         void AddExperience(ulong steamID, float experience);